first preview

+++ src/core/lib/task.ts
@@ -14,43 +14,55 @@ interface Itask {
   createdAt: string;
 }
 
+interface ItaskFirebase {
+  id: string;
+  title: string;
+  description: string;
+  status: TaskStatus;
+  createdAt: Timestamp;
+}
+
 class Task {
-  private data: Itask;
-  constructor(data: Itask) {
-    this.data = data;
-  }
-  public static async create({ title, description }:{ title: string; description: string;}) {
-    const newData: Itask = {
-      id: crypto.randomUUID(),
-      title: title,
-      description: description,
-      status: TaskStatus.IN_PROGRESS,
-      createdAt: new Date().toISOString(),
-    }
-    const docRef = doc(db, "tasks", newData.id);
-    /* await setDoc(docRef, {
+  private data: Itask = {
+    id: "",
+    createdAt: "",
+    description: "",
+    status: TaskStatus.IN_PROGRESS,
+    title: ""
+  };
+  constructor() {}
+  public async create({ title, description }:{ title: string; description: string;}) {
+    if (this.data.id) throw new Error("Already exist")
+    this.data.id = crypto.randomUUID()
+    this.data.title = title
+    this.data.description = description
+    this.data.status = TaskStatus.IN_PROGRESS
+    this.data.createdAt = new Date().toISOString()
+    const docRef = doc(db, "tasks", this.data.id);
+    await setDoc(docRef, {
       title: this.data.title,
       description: this.data.description,
       status: this.data.status,
       createdAt: Timestamp.fromDate(new Date(this.data.createdAt)),
-    }) */
-    return newData
+    })
   }
 
   public async update({title, description, status}: { title: string; description: string; status: TaskStatus}) {
+    if (!this.data.id) throw new Error("Element not exist")
     this.data.title = title;
     this.data.description = description;
     this.data.status = status;
-    /* await setDoc(docRef, {
+    const docRef = doc(db, "tasks", this.data.id);
+    await setDoc(docRef, {
       title: this.data.title,
       description: this.data.description,
       status: this.data.status,
       createdAt: Timestamp.fromDate(new Date(this.data.createdAt)),
-    }) */
-    return true
+    })
   }
 
   public async delete() {
+    if (!this.data.id) throw new Error("Element not exist")
     const docRef = doc(db, "tasks", this.data.id);
     await deleteDoc(docRef);
     return true
@@ -59,7 +71,11 @@ class Task {
   public get() {
     return this.data;
   }
+
+  public load(data: Itask){
+    this.data = data
+  }
 }
 
 export { TaskStatus, Task };
-export type { Itask };
+export type { Itask, ItaskFirebase };
